created release 4444/4446 from Development-branch

+++ plugins/qfe_helpeditor/qfehelpeditorwidget.cpp
@@ -517,28 +517,30 @@ void QFEHelpEditorWidget::setWinID(int winID)
 
 void QFEHelpEditorWidget::autosave()
 {
-    QDir d(QFPluginServices::getInstance()->getConfigFileDirectory());
-    QString ffn= d.absoluteFilePath(QString("helpeditor%1_autosave.html").arg(m_winID));
-    if (!getScript().isEmpty() && !qfFileEqualsString(ffn, ui->edtScript->getEditor()->toPlainText().toUtf8())) {
-        d.mkpath(d.absolutePath());
-        // copy _autosave_old.html --> _autosave_older.html
-        QString fn= d.absoluteFilePath(QString("helpeditor%1_autosave_old.html").arg(m_winID));
-        QString newfn= d.absoluteFilePath(QString("helpeditor%1_autosave_older.html").arg(m_winID));
-        if (QFile::exists(fn)) {
-            if (QFile::exists(newfn)) QFile::remove(newfn);
-            QFile::copy(fn, newfn);
-        }
+    if (isVisible()) {
+        QDir d(QFPluginServices::getInstance()->getConfigFileDirectory());
+        QString ffn= d.absoluteFilePath(QString("helpeditor%1_autosave.html").arg(m_winID));
+        if (!getScript().isEmpty() && !qfFileEqualsString(ffn, ui->edtScript->getEditor()->toPlainText().toUtf8())) {
+            d.mkpath(d.absolutePath());
+            // copy _autosave_old.html --> _autosave_older.html
+            QString fn= d.absoluteFilePath(QString("helpeditor%1_autosave_old.html").arg(m_winID));
+            QString newfn= d.absoluteFilePath(QString("helpeditor%1_autosave_older.html").arg(m_winID));
+            if (QFile::exists(fn)) {
+                if (QFile::exists(newfn)) QFile::remove(newfn);
+                QFile::copy(fn, newfn);
+            }
 
-        // copy _autosave.html --> _autosave_old.html
-        fn= ffn;
-        newfn= d.absoluteFilePath(QString("helpeditor%1_autosave_old.html").arg(m_winID));
-        if (QFile::exists(fn)) {
-            if (QFile::exists(newfn)) QFile::remove(newfn);
-            QFile::copy(fn, newfn);
+            // copy _autosave.html --> _autosave_old.html
+            fn= ffn;
+            newfn= d.absoluteFilePath(QString("helpeditor%1_autosave_old.html").arg(m_winID));
+            if (QFile::exists(fn)) {
+                if (QFile::exists(newfn)) QFile::remove(newfn);
+                QFile::copy(fn, newfn);
+            }
+            saveFile(fn, false);
+        } else {
+            qDebug()<<"QFEHelpEditorWidget::autosave(): file contents didn't change!";
         }
-        saveFile(fn, false);
-    } else {
-        qDebug()<<"QFEHelpEditorWidget::autosave(): file contents didn't change!";
     }
     QTimer::singleShot(AUTOSAVE_INTERVAL_MSEC, this, SLOT(autosave()));
 }
